Excepetion Handling and Bean Validation

+++ src/main/java/com/devsuperior/desafiocrudclientes/services/ClientService.java
@@ -9,6 +9,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.devsuperior.desafiocrudclientes.DTO.ClientDTO;
 import com.devsuperior.desafiocrudclientes.entities.Client;
 import com.devsuperior.desafiocrudclientes.repositories.ClientRepository;
+import com.devsuperior.desafiocrudclientes.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -18,7 +21,7 @@ public class ClientService {
 	
 	@Transactional(readOnly=true)
 	public ClientDTO findById(Long id) {
-		Client client = repository.findById(id).get();
+		Client client = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado!"));
 		return new ClientDTO(client);
 	}
 	
@@ -28,13 +31,14 @@ public class ClientService {
 		return client.map(x -> new ClientDTO(x));
 	}
 
+	@Transactional
 	public ClientDTO insert(ClientDTO dto) {
 		Client entity = new Client();
 		
 		entity.setName(dto.getName());
 		entity.setCpf(dto.getCpf());
 		entity.setIncome(dto.getIncome());
-		entity.setBithDate(dto.getBithDate());
+		entity.setBirthDate(dto.getBirthDate());
 		entity.setChildren(dto.getChildren());
 		
 		entity = repository.save(entity);
@@ -44,20 +48,29 @@ public class ClientService {
 	
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
+		try {
 		Client entity = repository.getReferenceById(id);
 		
 		entity.setName(dto.getName());
 		entity.setCpf(dto.getCpf());
 		entity.setIncome(dto.getIncome());
-		entity.setBithDate(dto.getBithDate());
+		entity.setBirthDate(dto.getBirthDate());
 		entity.setChildren(dto.getChildren());
 		
 		entity = repository.save(entity);
 		
 		return new ClientDTO(entity);
+		}
+		catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException ("Cliente não encontrado");
+		}
 	}
 	
+	@Transactional
 	public void delete (Long id) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Cliente não encontrato");
+		}
 		repository.deleteById(id);
 	}
 }
